+++ util/prometheusUtil/prometheusClient.go
@@ -1,4 +1,4 @@
-package util
+package prometheusUtil
 
 import (
 	"context"
