+++ util/prometheusUtil/prometheusClient_test.go
@@ -1,4 +1,4 @@
-package util
+package prometheusUtil
 
 import (
 	"context"
